feat: :sparkles: Metrics Snapshot workflow
