+++ package.json
@@ -23,7 +23,8 @@
     "prune-branches": "git removed-branches --prune --force",
     "test:upload": "npx tsx scripts/upload-test-posters.ts",
     "test:dummy": "npx tsx scripts/create-dummy-posters.ts",
-    "scripts:add:posters": "npx tsx scripts/add-extracted-posters.ts"
+    "scripts:add:posters": "npx tsx scripts/add-extracted-posters.ts",
+    "metrics:snapshot": "npx tsx scripts/compute-metrics-snapshot.ts"
   },
   "dependencies": {
     "@inspira-ui/plugins": "0.0.2",
